+++ mstring.h
@@ -10,8 +10,8 @@
 
 typedef struct{
     char* chars;
-    uint32_t length;
-    uint32_t blocks;
+    size_t length;
+    size_t blocks;
 }Mstring;
 
 void free_string(Mstring* str); // changed from string_dispose() to free_mstring() to be more compatible with the other free methods (see Mexecution.h/c)
@@ -20,30 +20,32 @@ void free_string(Mstring* str); // changed from string_dispose() to free_mstring
 Mstring* __string();
 Mstring* _getString(const char* const s); // convenient constructor
 // copying
-Mstring* _stringCopy(Mstring* const src,uint32_t length);
+Mstring* _stringCopy(Mstring* const src,size_t length);
 
 bool string_empty(const Mstring* const str);
-uint32_t string_length(const Mstring* const str);
-Mstring* string_setlength(Mstring* const str, uint32_t length); // MDH@26FEB2018: we might want to set the length (to a smaller one)
+
+size_t string_length(const Mstring* const str);
+Mstring* string_setlength(Mstring* const str, size_t length); // MDH@26FEB2018: we might want to set the length (to a smaller one)
+
 void string_synclength(Mstring* const str); // MDH@02JUN2019: check the length (if a \0 is in front of the current length)
 
-char string_char(const Mstring* const str,uint32_t pos);
+char string_char(const Mstring* const str,size_t pos);
 char string_last_char(const Mstring* const str);
-char* string_remainder(Mstring* const str,uint32_t pos);
+char* string_remainder(Mstring* const str,size_t pos);
 char* string(Mstring* const str);
 
-char* _stringstart(const Mstring* const str,uint32_t length); // returns a copy of the first part of the string
+char* _stringstart(const Mstring* const str,size_t length); // returns a copy of the first part of the string
 
-bool string_shorten(Mstring* const str,uint32_t length);
+bool string_shorten(Mstring* const str,size_t length);
 
 int32_t string_find(const Mstring* const str,char c);
 
 // changing the string
-char string_removed_char(Mstring* const str,uint32_t pos);
+char string_removed_char(Mstring* const str,size_t pos);
 
-Mstring* string_insert_char(Mstring* const str,uint32_t pos,char c);
+Mstring* string_insert_char(Mstring* const str,size_t pos,char c);
 Mstring* string_append_char(Mstring* const str,char c);
-Mstring* string_setchar(Mstring* const str,char c,uint32_t pos);
+Mstring* string_setchar(Mstring* const str,char c,size_t pos);
 
 // MDH@26FEB2019: can we append a text as a whole???
 Mstring* string_append(Mstring* const str,const char* pc);
@@ -52,5 +54,5 @@ Mstring* string_prepend(Mstring* const str,const char* pc);
 void string_reverse(Mstring* const str);
 
 Mstring* string_append_chars(Mstring* const str,const char* pc,size_t count); // MDH@24SEP2019: we need to be able to append count characters from pc
-char string_replacedchar(Mstring* const str,char c,uint32_t pos); // returns the character at position pos replaced by c (but does not change the length ever)
+char string_replacedchar(Mstring* const str,char c,size_t pos); // returns the character at position pos replaced by c (but does not change the length ever)
 #endif /* __MSTRING_H__ */
